Some cleanup before moving Keystore PAssword to HashiCorp
