+++ cryptoservice/src/main/java/com/projectwork/cryptoservice/CryptoserviceApplication.java
@@ -26,3 +26,12 @@ public class CryptoserviceApplication {
 
 }
 
+/**
+ * SecureCodingPractices later implementation:
+ * - OWASP [107] In catch-Blöcken keine Stacktraces direkt loggen -> später generische Fehler-Response + Logging verbessern
+ * - OWASP [131] Least Privilege: Sicherstellen, dass der Service nur mit nötigen Dateisystem-Rechten läuft (kein root)
+ * - OWASP [105] Prüfen, ob du ein FIPS 140-2 konformes JCE (Java Cryptography Extension) nutzt -> gut für Doku
+ * - OWASP [142] Datei- und Zugriffsrechte auf den Keystore beschränken (nur les- und schreibbar für Service-User)
+ * - OWASP [143] Später TLS -> Communication Security sicherstellen
+ */
+
