+++ cryptoservice/src/main/resources/application.properties
@@ -1,2 +1,1 @@
 spring.application.name=cryptoservice
-keystore.password=${KEYSTORE_PASSWORD}
